requirements.txt
flask
gunicorn
python-dotenv
trading-ig
pydantic
tenacity
